[proj4] Got the average to work and use insert to add Students to the list. Passes the first test

+++ proj4/student.h
@@ -5,13 +5,13 @@ typedef struct student_struct
 {
     char name[16];
     int age;
-    double gpa;
+    float gpa;
     struct student_struct *next;
 } Student;
 
 Student head; //The head of the linked list
 
-Student *makeStudent(char name[16], int age, double gpa);
+Student *makeStudent(char name[16], int age, float gpa);
 void push(Student *student);
 Student *findTopStudent();
 float getAverageGPA();
